requirements.txt
fastapi==0.73.0
propelauth-fastapi==1.1.0
python-dotenv==0.19.2
uvicorn==0.17.0
